% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{addCodelist}
\alias{addCodelist}
\title{Add a codelist to a SAP}
\usage{
addCodelist(sap, codelist, validate = TRUE)
}
\arguments{
\item{sap}{A \code{sap} object.}

\item{codelist}{A \code{sap_codelist} object.}

\item{validate}{Whether to validate the whole SAP after the change. An id
already in use is refused either way.}
}
\value{
The updated \code{sap} object.
}
\description{
Add a codelist to a SAP
}
