% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{createSap}
\alias{createSap}
\title{Create a complete SAP from its components}
\usage{
createSap(
  study,
  dataSources = list(),
  dataSourceModifications = list(),
  codelists = list(),
  cohorts = list(),
  analyses = list(),
  validate = TRUE
)
}
\arguments{
\item{study}{A \code{sap_study} object.}

\item{dataSources}{List of \code{sap_data_source} objects.}

\item{dataSourceModifications}{List of \code{sap_data_source_modification} objects.}

\item{codelists}{List of \code{sap_codelist} objects.}

\item{cohorts}{List of \code{sap_cohort} objects.}

\item{analyses}{List of \code{sap_analysis} objects.}

\item{validate}{Whether to validate the complete SAP. With \code{FALSE} the
document is assembled from drafts and \code{\link[=checkSap]{checkSap()}} reports its problems.}
}
\value{
An object of class \code{sap}.
}
\description{
Create a complete SAP from its components
}
