% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{getSapComponent}
\alias{getSapComponent}
\title{Get a component from a SAP by id}
\usage{
getSapComponent(sap, collection, id)
}
\arguments{
\item{sap}{A \code{sap} object.}

\item{collection}{One of the SAP collection names.}

\item{id}{The component id.}
}
\value{
The component.
}
\description{
Get a component from a SAP by id
}
