% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{newSap}
\alias{newSap}
\title{Create a structured Statistical Analysis Plan}
\usage{
newSap(x, validate = TRUE)
}
\arguments{
\item{x}{A named R list containing the SAP fields.}

\item{validate}{Whether to validate the complete SAP before returning it.}
}
\value{
An object of class \code{sap}.
}
\description{
Create a structured Statistical Analysis Plan
}
