% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{sapComponentIds}
\alias{sapComponentIds}
\title{List the ids in a SAP collection}
\usage{
sapComponentIds(sap, collection)
}
\arguments{
\item{sap}{A \code{sap} object.}

\item{collection}{One of the SAP collection names.}
}
\value{
A character vector of ids.
}
\description{
List the ids in a SAP collection
}
