% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_schema.R
\name{sapSchemaVersion}
\alias{sapSchemaVersion}
\title{The current SAP schema version}
\usage{
sapSchemaVersion()
}
\value{
A single string, e.g. \code{"0.1.0"}.
}
\description{
The current SAP schema version
}
