% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{validateSap}
\alias{validateSap}
\title{Validate a SAP, throwing an error when problems are found}
\usage{
validateSap(x)
}
\arguments{
\item{x}{A SAP object or named list.}
}
\value{
The validated SAP, invisibly.
}
\description{
Validate a SAP, throwing an error when problems are found
}
