% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{addCodelist}
\alias{addCodelist}
\title{Add a codelist to a SAP}
\usage{
addCodelist(sap, codelist)
}
\arguments{
\item{sap}{A \code{sap} object.}

\item{codelist}{A \code{sap_codelist} object.}
}
\value{
The updated \code{sap} object.
}
\description{
Add a codelist to a SAP
}
