% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{addCohort}
\alias{addCohort}
\title{Add a cohort to a SAP}
\usage{
addCohort(sap, cohort)
}
\arguments{
\item{sap}{A \code{sap} object.}

\item{cohort}{A \code{sap_cohort} object.}
}
\value{
The updated \code{sap} object.
}
\description{
Add a cohort to a SAP
}
