% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{addDataSource}
\alias{addDataSource}
\title{Add a data source to a SAP}
\usage{
addDataSource(sap, dataSource)
}
\arguments{
\item{sap}{A \code{sap} object.}

\item{dataSource}{A \code{sap_data_source} object.}
}
\value{
The updated \code{sap} object.
}
\description{
Add a data source to a SAP
}
