% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{addDataSourceModification}
\alias{addDataSourceModification}
\title{Add a data-source modification to a SAP}
\usage{
addDataSourceModification(sap, dataSourceModification)
}
\arguments{
\item{sap}{A \code{sap} object.}

\item{dataSourceModification}{A \code{sap_data_source_modification} object.}
}
\value{
The updated \code{sap} object.
}
\description{
Add a data-source modification to a SAP
}
