% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{addDataSourceModification}
\alias{addDataSourceModification}
\title{Add a data-source modification to a SAP}
\usage{
addDataSourceModification(sap, dataSourceModification)
}
\description{
Add a data-source modification to a SAP
}
