% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{addSapComponent}
\alias{addSapComponent}
\title{Add a component to a SAP}
\usage{
addSapComponent(sap, component)
}
\arguments{
\item{sap}{A \code{sap} object.}

\item{component}{A \verb{sap_*} component object.}
}
\value{
The updated \code{sap} object.
}
\description{
Add a component to a SAP
}
