% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{createSap}
\alias{createSap}
\title{Create a complete SAP from its components}
\usage{
createSap(
  study,
  dataSources = list(),
  dataSourceModifications = list(),
  codelists = list(),
  cohorts = list(),
  analyses = list()
)
}
\description{
Create a complete SAP from its components
}
