% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{newSap}
\alias{newSap}
\title{Create an OMOP Standardised Analysis Plan}
\usage{
newSap(x)
}
\arguments{
\item{x}{An R list to create a \code{sap} object.}
}
\value{
A sap object.
}
\description{
Create an OMOP Standardised Analysis Plan
}
