% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{newSapCodelist}
\alias{newSapCodelist}
\title{Create a SAP codelist component from an external codelist object}
\usage{
newSapCodelist(id, name, type, content)
}
\description{
Create a SAP codelist component from an external codelist object
}
