% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{newSapCohort}
\alias{newSapCohort}
\title{Create a SAP cohort component}
\usage{
newSapCohort(id, name, dataSourceId, type, parameters = list())
}
\arguments{
\item{id}{Immutable cohort identifier.}

\item{name}{Display name of the cohort.}

\item{dataSourceId}{Identifier of the data source(s) used by the cohort.}

\item{type}{Cohort type.}

\item{parameters}{Type-specific cohort parameters.}
}
\value{
An object of class \code{sap_cohort}.
}
\description{
Create a SAP cohort component
}
