% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{newSapDataSourceModification}
\alias{newSapDataSourceModification}
\title{Create a data-source modification component}
\usage{
newSapDataSourceModification(id, name, type, dataSourceId, parameters = list())
}
\arguments{
\item{id}{Immutable modification identifier.}

\item{name}{Display name of the modification.}

\item{type}{Data-source modification type.}

\item{dataSourceId}{Identifier of the affected data source(s).}

\item{parameters}{Type-specific modification parameters.}
}
\value{
An object of class \code{sap_data_source_modification}.
}
\description{
Create a data-source modification component
}
