% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{newSapDataSourceModification}
\alias{newSapDataSourceModification}
\title{Create a data-source modification component}
\usage{
newSapDataSourceModification(id, name, type, dataSourceId, parameters = list())
}
\description{
Create a data-source modification component
}
