% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{newSapStudy}
\alias{newSapStudy}
\title{Create the study metadata component of a SAP}
\usage{
newSapStudy(
  studyId,
  title,
  authors = character(),
  version = "v1.0.0",
  description = NULL
)
}
\description{
Create the study metadata component of a SAP
}
