% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{newSapStudy}
\alias{newSapStudy}
\title{Create the study metadata component of a SAP}
\usage{
newSapStudy(
  studyId,
  title,
  authors = character(),
  version = "v1.0.0",
  description = NULL
)
}
\arguments{
\item{studyId}{Immutable study identifier.}

\item{title}{Study title.}

\item{authors}{Character vector of study authors.}

\item{version}{Study version label.}

\item{description}{Optional study description.}
}
\value{
An object of class \code{sap_study}.
}
\description{
Create the study metadata component of a SAP
}
