% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap.R
\name{updateStudy}
\alias{updateStudy}
\title{Replace the study metadata of a SAP}
\usage{
updateStudy(sap, study)
}
\arguments{
\item{sap}{A \code{sap} object.}

\item{study}{A \code{sap_study} object.}
}
\value{
The updated \code{sap} object.
}
\description{
Replace the study metadata of a SAP
}
